build(pkgs): Implement build-status infrastructure
Why this change is needed?
==========================

The test matrix { package x version x system } contains various elements with
problems and the sheer size of the matrix makes it difficult to manage (e.g. to
know which CI jobs to mark as allowed to fail, to know on which elements the
checkPhase can be run or which tests need to be removed.

Investigation of the problems shows that many of the failures are caused by a
handful of tests that fail on a specific version and system. Instead of marking
the whole job as failed, it is better to omit the failing tests, as the rest of
the suite (more than 99%!) will still pass.

What does this change do?
=========================

For each package:
* Add a `pkgs/${packageName}/build-status.json` file containing
  * each version known to have problems
    * each system on which this version is known to have problems
      * `build` - does it build succesfully
        * if `false` the CI will be marked as allowed to fail
      * `check` - does the check phase complete
      * `skippedTests` - any tests that are necessary to be skipped to allow the
        checkPhase to succeed
* Add to its derivation `passthru.buildStatus` field surfacing this information
  on the Nix eval level
* Set `doCheck` to `buildStatus.check`
* Add to the `checkPhase`
  * for each `test_file` in buildStatus.skippedTests
    * a line `rm -v ${test_file}` that will remove the test on the corresponding
      tripple (package, version, system)